moved FunctorFilter to core
