+++ modules/aql/arrow-query-language/src/main/kotlin/arrow/aql/extensions/try.kt
@@ -15,7 +15,7 @@ import arrow.core.extensions.`try`.applicative.applicative
 import arrow.core.extensions.`try`.foldable.foldable
 import arrow.core.extensions.`try`.functor.functor
 import arrow.mtl.extensions.`try`.functorFilter.functorFilter
-import arrow.mtl.typeclasses.FunctorFilter
+import arrow.typeclasses.FunctorFilter
 import arrow.typeclasses.Applicative
 import arrow.typeclasses.Foldable
 import arrow.typeclasses.Functor
