+++ modules/core/arrow-core-data/src/main/kotlin/arrow/typeclasses/FunctorFilter.kt
@@ -1,4 +1,4 @@
-package arrow.mtl.typeclasses
+package arrow.typeclasses
 
 import arrow.Kind
 import arrow.core.None
@@ -7,10 +7,9 @@ import arrow.core.PartialFunction
 import arrow.core.Some
 import arrow.core.identity
 import arrow.core.lift
-import arrow.typeclasses.Functor
 
 /**
- * ank_macro_hierarchy(arrow.mtl.typeclasses.FunctorFilter)
+ * ank_macro_hierarchy(arrow.typeclasses.FunctorFilter)
  *
  * A Functor with the ability to [mapFilter].
  * Enables [collect] based on [PartialFunction] predicates.
