+++ modules/docs/arrow-docs/src/main/kotlin/AQLHelper.kt
@@ -6,7 +6,7 @@ import arrow.aql.box.functorFilter.functorFilter
 import arrow.core.Eval
 import arrow.core.Option
 import arrow.extension
-import arrow.mtl.typeclasses.FunctorFilter
+import arrow.typeclasses.FunctorFilter
 import arrow.typeclasses.Foldable
 import arrow.typeclasses.Functor
 import arrow.undocumented
