+++ modules/mtl/arrow-mtl/src/main/kotlin/arrow/mtl/extensions/TryFunctorFilter.kt
@@ -6,7 +6,7 @@ import arrow.core.Option
 import arrow.core.Try
 import arrow.core.fix
 import arrow.extension
-import arrow.mtl.typeclasses.FunctorFilter
+import arrow.typeclasses.FunctorFilter
 
 @extension
 interface TryFunctorFilter : FunctorFilter<ForTry> {
