+++ modules/mtl/arrow-mtl/src/main/kotlin/arrow/mtl/extensions/listk.kt
@@ -9,7 +9,7 @@ import arrow.core.ListK
 import arrow.core.ListKOf
 import arrow.core.fix
 import arrow.extension
-import arrow.mtl.typeclasses.FunctorFilter
+import arrow.typeclasses.FunctorFilter
 import arrow.mtl.typeclasses.MonadCombine
 import arrow.mtl.typeclasses.MonadFilter
 import arrow.core.combineK as listCombineK
