+++ modules/mtl/arrow-mtl/src/main/kotlin/arrow/mtl/extensions/option.kt
@@ -10,7 +10,7 @@ import arrow.core.Tuple2
 import arrow.core.fix
 import arrow.core.orElse
 import arrow.extension
-import arrow.mtl.typeclasses.FunctorFilter
+import arrow.typeclasses.FunctorFilter
 import arrow.mtl.typeclasses.MonadCombine
 import arrow.mtl.typeclasses.MonadFilter
 import arrow.mtl.typeclasses.TraverseFilter
